+++ packages/orq-rc/src/models/operations/createchatcompletion.ts
@@ -1229,7 +1229,7 @@ export type FilterBy1 =
   | CreateChatCompletion1Nin;
 
 /**
- * The metadata filter to apply to the search. Check the [Searching a Knowledge Base](https://docs.orq.ai/docs/knowledge/api#knowledge-base-search) for more information.
+ * The metadata filter to apply to the search. Check the [Searching a Knowledge Base](https://docs.orq.ai/docs/ai-studio/ai-engineering/knowledge-bases#search-a-knowledge-base) for more information.
  */
 export type FilterBy = FilterByAnd | FilterByOr | {
   [k: string]:
@@ -1266,7 +1266,7 @@ export type SearchOptions = {
  */
 export type RerankConfig = {
   /**
-   * The name of the rerank model to use. Refer to the [model list](https://docs.orq.ai/docs/proxy#/rerank-models).
+   * The name of the rerank model to use. Refer to the [model list](https://docs.orq.ai/docs/ai-gateway/supported-models#rerank-models).
    */
   model: string;
   /**
@@ -1284,7 +1284,7 @@ export type RerankConfig = {
  */
 export type AgenticRagConfig = {
   /**
-   * The name of the model for the Agent to use. Refer to the [model list](https://docs.orq.ai/docs/proxy#/chat-models).
+   * The name of the model for the Agent to use. Refer to the [model list](https://docs.orq.ai/docs/ai-gateway/supported-models#chat-models).
    */
   model: string;
 };
@@ -1303,7 +1303,7 @@ export type CreateChatCompletionKnowledgeBases = {
    */
   searchType?: SearchType | null | undefined;
   /**
-   * The metadata filter to apply to the search. Check the [Searching a Knowledge Base](https://docs.orq.ai/docs/knowledge/api#knowledge-base-search) for more information.
+   * The metadata filter to apply to the search. Check the [Searching a Knowledge Base](https://docs.orq.ai/docs/ai-studio/ai-engineering/knowledge-bases#search-a-knowledge-base) for more information.
    */
   filterBy?: FilterByAnd | FilterByOr | {
     [k: string]: