+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M28ES586K1MYDKVVXKZ5JA9V"),
+  id: z.string().default("01M2EYE91EZ299XBV52D99YSA5"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>