+++ packages/orq-rc/RELEASES.md
@@ -12299,3 +12299,13 @@ Based on:
 - [typescript v4.15.0-rc.41] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.41 - packages/orq-rc
+
+## 2026-09-14 03:15:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.42] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.42 - packages/orq-rc